VP-1310: Migrate to V3
- Structure changed;
- Adopted code;
- Settings addition in process;
