+++ .nuke
@@ -1,0 +1,1 @@
+VirtoCommerce.AuthorizeNet.sln
