--- Authorize.Net/packages.config
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AuthorizeNet" version="1.8.10" targetFramework="net451" />
-  <package id="Common.Logging" version="3.1.0" targetFramework="net451" />
-  <package id="Common.Logging.Core" version="3.1.0" targetFramework="net451" />
-  <package id="CommonServiceLocator" version="1.3" targetFramework="net451" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.4" targetFramework="net461" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.4" targetFramework="net461" />
-  <package id="MSBuildTasks" version="1.5.0.235" targetFramework="net461" developmentDependency="true" />
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
-  <package id="Unity" version="4.0.1" targetFramework="net451" />
-  <package id="VirtoCommerce.Domain" version="2.25.31" targetFramework="net461" />
-  <package id="VirtoCommerce.Module" version="0.18.0" targetFramework="net461" developmentDependency="true" />
-  <package id="VirtoCommerce.Platform.Core" version="2.13.42" targetFramework="net461" />
-</packages>
