+++ artifacts/VirtoCommerce.AuthorizeNet3.0.0-rc.3/Content/Authorizenet_logo.png
