+++ src/VirtoCommerce.AuthorizeNet.Web/VirtoCommerce.AuthorizeNet.Web.csproj
@@ -1,0 +1,21 @@
+﻿ <Project Sdk="Microsoft.NET.Sdk.Web">
+ 	<PropertyGroup>
+ 		<TargetFramework>netcoreapp3.1</TargetFramework>
+ 		<ApplicationIcon />
+ 		<OutputType>Library</OutputType>
+        <noWarn>1591</noWarn>
+        <IsPackable>false</IsPackable>
+ 	</PropertyGroup>
+	
+	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <NoWarn>1701;1702;1705;1591;</NoWarn>
+    </PropertyGroup>
+	
+	<ItemGroup>
+	  <PackageReference Include="AuthorizeNet.NetStandard" Version="0.1.0" />
+	  <PackageReference Include="VirtoCommerce.OrdersModule.Core" Version="3.0.0-rc.3-001" />
+	  <PackageReference Include="VirtoCommerce.PaymentModule.Core" Version="3.0.0-rc.3-001" />
+	  <PackageReference Include="VirtoCommerce.Platform.Core" Version="3.0.0-rc3-0141" />
+	  <PackageReference Include="VirtoCommerce.StoreModule.Core" Version="3.0.0-rc.3" />
+	</ItemGroup>
+ </Project>
