+++ src/VirtoCommerce.AuthorizeNet.Web/module.manifest
@@ -1,0 +1,87 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<module>
+    <id>VirtoCommerce.AuthorizeNet</id>
+    <version>3.0.0</version>
+	  <version-tag>rc.3</version-tag>
+    <platformVersion>3.0.0</platformVersion>
+    <dependencies>
+      <dependency id="VirtoCommerce.Store" version="3.0.0" />
+			<dependency id="VirtoCommerce.Orders" version="3.0.0-rc.3-001" />
+			<dependency id="VirtoCommerce.Payment" version="3.0.0-rc.3-001" />
+		</dependencies>
+
+    <title>Authorize.Net DPM payment gateway</title>
+    <description>APIs and UI for Authorize.Net Direct Post Method (DPM) payment gateway</description>
+    <authors>
+        <author>Virto Commerce</author>
+    </authors>
+    <owners>
+        <owner>Virto Commerce</owner>
+    </owners>
+    <projectUrl>https://virtocommerce.com/apps/extensions/authorizenet-dpm-payment-gateway</projectUrl>
+    <iconUrl>https://raw.githubusercontent.com/VirtoCommerce/vc-module-Authorize.Net/master/Authorize.Net/Content/Authorizenet_logo.png</iconUrl>
+    <releaseNotes>First version.</releaseNotes>
+    <copyright>Copyright © 2011-2020 Virto Commerce. All rights reserved</copyright>
+    <tags>Authorize.Net payment gateway</tags>
+
+    <assemblyFile>VirtoCommerce.AuthorizeNet.Web.dll</assemblyFile>
+    <moduleType>VirtoCommerce.AuthorizeNet.Web.Module, VirtoCommerce.AuthorizeNet.Web</moduleType>
+
+    <settings>
+        <group name="Payment methods|Authorize.Net|Credentials">
+            <setting>
+                <name>AuthorizeNet.ApiLogin</name>
+                <valueType>string</valueType>
+                <title>API Login ID</title>
+                <description>API Login ID from Authorize.Net credentials</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.TxnKey</name>
+                <valueType>securestring</valueType>
+                <title>Transaction Key</title>
+                <description>Transaction Key from Authorize.Net credentials</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.Mode</name>
+                <valueType>string</valueType>
+                <defaultValue>test</defaultValue>
+                <allowedValues>
+                    <value>test</value>
+                    <value>real</value>
+                </allowedValues>
+                <title>Mode</title>
+                <description>Mode of Authorize.Net payment gateway (test or real)</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.ConfirmationUrl</name>
+                <valueType>string</valueType>
+                <defaultValue>{VC manager URL}/api/payments/an/registerpayment</defaultValue>
+                <title>Confirmation URL</title>
+                <description>URL for payment confirmation in VC Manager API. {VC manager URL}/api/payments/an/registerpayment</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.ThankYouPageRelativeUrl</name>
+                <defaultValue>cart/thanks</defaultValue>
+                <title>Thank you page relative URL</title>
+                <description>Storefront thank you page relative URL</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.PaymentActionType</name>
+                <valueType>string</valueType>
+                <allowedValues>
+                    <value>Authorization/Capture</value>
+                    <value>Sale</value>
+                </allowedValues>
+                <defaultValue>Authorization/Capture</defaultValue>
+                <title>Payment action type</title>
+                <description>Action type of payment</description>
+            </setting>
+            <setting>
+                <name>AuthorizeNet.SHA2Hash</name>
+                <valueType>securestring</valueType>
+                <title>SHA2 hash key</title>
+                <description>Authorize Net SHA2 hash key used for relay response validation</description>
+            </setting>
+        </group>
+    </settings>
+</module>
